Changes needed to get it all compiling on Windows without warnings.
This included changing references to snprintf so that MSVC would be happy.  It also
included a fix to socket_destroy() to use closesocket() instead of close() on Windows.

+++ src/examples/tag_rw.c
@@ -158,7 +158,7 @@ int main(int argc, char **argv)
     real32_t f_val;
     int i;
     int rc;
-    int timeout;
+    int64_t timeout;
 
     parse_args(argc, argv);
